Update info on creating DataProc cluster
When creating DataProc clusters in the new GCP organization (ssb.no),
it has to be done slightly differently from in the ssbmod.net
organization.
Included a comment on network setup, as the new GCP-org doesn't have
(or allow auto-creation of) default networks. The previous version
of command example used `default` subnet, which doesn't work in the
new setup/org. (The necessary network setup has to be done by the
Stratus platform team.)